+++ src/components/cities-list/cities-list.jsx
@@ -34,8 +34,7 @@ class CitiesList extends React.PureComponent {
 
 CitiesList.propTypes = {
   cities: PropTypes.arrayOf(PropTypes.string).isRequired,
-  changeCity: PropTypes.func,
-  onChangeCity: PropTypes.func,
+  onChangeCity: PropTypes.func.isRequired,
 };
 
 export default CitiesList;
